make review changes
