+++ src/loaders.js
@@ -1,3 +1,7 @@
+import loadAbout from './about';
+import loadMenu from './menu';
+import loadContact from './contact';
+
 export default function loadPage() {
   const content = document.getElementById('content');
   const h1 = document.createElement('h1');
@@ -45,44 +49,6 @@ export default function loadPage() {
   content.appendChild(aboutTab);
   content.appendChild(menuTab);
   content.appendChild(contactTab);
-
-  function loadAbout() {
-    const menuItemTitle = document.createElement('h3');
-    const menuItemBody = document.createElement('p');
-    const tabcontent = document.getElementById('aboutus');
-    menuItemTitle.textContent = 'About Us';
-    menuItemBody.textContent = 'We are a small start-up restaurant '
-      + 'offering convenient and affordable meals to our clients';
-    tabcontent.appendChild(menuItemTitle);
-    tabcontent.appendChild(menuItemBody);
-  }
-
-  function loadMenu() {
-    const menuItemTitle = document.createElement('h3');
-    const menuItemBody = document.createElement('p');
-    const tabcontent = document.getElementById('menu');
-    menuItemTitle.textContent = 'Menu';
-    menuItemBody.textContent = 'BREAKFAST\r\nToast, bacon, sausages, eggs, black pudding,'
-      + 'baked beans, tomatoes, mushrooms and a mug of coffee or tea.\r\n\r\n'
-      + 'LUNCH\r\nMouthwatering chunks of Certified Angus Beef brisket, juicy pork or chunks of delicious '
-      + 'polish sausage served with coleslaw, French fries, Texas toast and pickles.\n\r\n'
-      + 'DINNER\r\nGrilled Chicken. Served with leftover brown rice seasoned with salt, pepper, '
-      + 'and cilantro. Topped with sliced avocado.';
-    tabcontent.appendChild(menuItemTitle);
-    tabcontent.appendChild(menuItemBody);
-    tabcontent.style.display = 'none';
-  }
-
-  function loadContact() {
-    const menuItemTitle = document.createElement('h3');
-    const menuItemBody = document.createElement('p');
-    const tabcontent = document.getElementById('contactus');
-    menuItemTitle.textContent = 'Contact Us';
-    menuItemBody.textContent = 'We are located at Plainsview Plaza off Mombasa Road in Nairobi.';
-    tabcontent.appendChild(menuItemTitle);
-    tabcontent.appendChild(menuItemBody);
-    tabcontent.style.display = 'none';
-  }
 
   loadAbout();
   loadMenu();
